Refine `Store` interface
